homework progress

+++ method/geometry calculator/Program.cs
@@ -10,7 +10,7 @@ namespace cube_properties
     {
         static void Main(string[] args)
         {
-            int n = int.Parse(Console.ReadLine());
+            double n = double.Parse(Console.ReadLine());
             string formulaSearch = Console.ReadLine().ToLower();
             double result = 0;
             if (formulaSearch == "face")
@@ -33,22 +33,22 @@ namespace cube_properties
             Console.WriteLine("{0:f2}", result);
         }
 
-        private static double GetVolume(int n)
+        private static double GetVolume(double n)
         {
             return Math.Pow(n, 3);
         }
 
-        private static double GetArea(int n)
+        private static double GetArea(double n)
         {
             return Math.Pow(n, 2) * 6;
         }
 
-        private static double GetSpaceDiagonals(int n)
+        private static double GetSpaceDiagonals(double n)
         {
             return Math.Sqrt(Math.Pow(n, 2) * 3);
         }
 
-        private static double GetFaceDiagonals(int n)
+        private static double GetFaceDiagonals(double n)
         {
             return Math.Sqrt(Math.Pow(n, 2) * 2);
         }
